refactor model interface
