+++ models/two_player_gan_model.py
@@ -16,15 +16,13 @@ You need to implement the following functions:
     <optimize_parameters>: Update network weights; it will be called in every training iteration.
 """
 import torch
-import numpy as np 
+import numpy as np
 from .base_model import BaseModel
-from . import networks
-from util.util import prepare_z_y, one_hot, visualize_imgs 
+from networks import networks
+from networks.loss import GANLoss, cal_gradient_penalty
+from util.util import one_hot
 from torch.distributions import Categorical
-from collections import OrderedDict
-from TTUR import fid
-from util.inception import get_inception_score
-from inception_pytorch import inception_utils
+
 
 class TwoPlayerGANModel(BaseModel):
     @staticmethod
@@ -38,11 +36,13 @@ class TwoPlayerGANModel(BaseModel):
         Returns:
             the modified parser.
         """
-        #parser.set_defaults(dataset_mode='aligned')  # You can rewrite default values for this model. For example, this model usually uses aligned dataset as its dataset
+        # parser.set_defaults(dataset_mode='aligned')  # You can rewrite default values for this model. For example, this model usually uses aligned dataset as its dataset
         if is_train:
-            parser.add_argument('--g_loss_mode', type=str, default='lsgan', help='lsgan | nsgan | vanilla | wgan | hinge | rsgan') 
-            parser.add_argument('--d_loss_mode', type=str, default='lsgan', help='lsgan | nsgan | vanilla | wgan | hinge | rsgan') 
-            parser.add_argument('--which_D', type=str, default='S', help='Standard(S) | Relativistic_average (Ra)') 
+            parser.add_argument('--g_loss_mode', type=str, default='lsgan',
+                                help='lsgan | nsgan | vanilla | wgan | hinge | rsgan')
+            parser.add_argument('--d_loss_mode', type=str, default='lsgan',
+                                help='lsgan | nsgan | vanilla | wgan | hinge | rsgan')
+            parser.add_argument('--which_D', type=str, default='S', help='Standard(S) | Relativistic_average (Ra)')
 
         return parser
 
@@ -71,48 +71,41 @@ class TwoPlayerGANModel(BaseModel):
             self.model_names = ['G']
 
         if self.opt.cgan:
-            probs = np.ones(self.opt.cat_num)/self.opt.cat_num 
+            probs = np.ones(self.opt.cat_num) / self.opt.cat_num
             self.CatDis = Categorical(torch.tensor(probs))
         # define networks 
-        self.netG = networks.define_G(opt.z_dim, opt.output_nc, opt.ngf, opt.netG,
-                opt.g_norm, opt.cgan, opt.cat_num, not opt.no_dropout, opt.init_type, opt.init_gain, self.gpu_ids)
-
-        if self.isTrain:  # define a discriminator; conditional GANs need to take both input and output images; Therefore, #channels for D is input_nc + output_nc
-            self.netD = networks.define_D(opt.input_nc, opt.ndf, opt.netD,
-                                          opt.d_norm, opt.cgan, opt.cat_num, opt.init_type, opt.init_gain, self.gpu_ids)
-
+        self.netG = networks.define_G(opt, self.device)
+        if self.isTrain:
+            # define a discriminator;
+            # conditional GANs need to take both input and output images;
+            # Therefore, #channels for D is input_nc + output_nc
+            self.netD = networks.define_D(opt, self.device)
         if self.isTrain:  # only defined during training time
             # define loss functions
-            self.criterionG = networks.GANLoss(opt.g_loss_mode, 'G', opt.which_D).to(self.device)
-            self.criterionD = networks.GANLoss(opt.d_loss_mode, 'D', opt.which_D).to(self.device)
+            self.criterionG = GANLoss(opt.g_loss_mode, 'G', opt.which_D).to(self.device)
+            self.criterionD = GANLoss(opt.d_loss_mode, 'D', opt.which_D).to(self.device)
             # initialize optimizers
             self.optimizer_G = torch.optim.Adam(self.netG.parameters(), lr=opt.lr_g, betas=(opt.beta1, opt.beta2))
             self.optimizer_D = torch.optim.Adam(self.netD.parameters(), lr=opt.lr_d, betas=(opt.beta1, opt.beta2))
             self.optimizers.append(self.optimizer_G)
             self.optimizers.append(self.optimizer_D)
-        # visulize settings 
-        self.N =int(np.trunc(np.sqrt(min(opt.batch_size, 64))))
-        if self.opt.z_type == 'Gaussian': 
-            self.z_fixed = torch.randn(self.N*self.N, opt.z_dim, 1, 1, device=self.device) 
-        elif self.opt.z_type == 'Uniform': 
-            self.z_fixed = torch.rand(self.N*self.N, opt.z_dim, 1, 1, device=self.device)*2. - 1. 
+        # visualize settings
+        self.N = int(np.trunc(np.sqrt(min(opt.batch_size, 64))))
+        if self.opt.z_type == 'Gaussian':
+            self.z_fixed = torch.randn(self.N * self.N, opt.z_dim, 1, 1, device=self.device)
+        elif self.opt.z_type == 'Uniform':
+            self.z_fixed = torch.rand(self.N * self.N, opt.z_dim, 1, 1, device=self.device) * 2. - 1.
         if self.opt.cgan:
-            yf = self.CatDis.sample([self.N*self.N])
-            self.y_fixed = one_hot(yf, [self.N*self.N, self.opt.cat_num])
+            yf = self.CatDis.sample([self.N * self.N])
+            self.y_fixed = one_hot(yf, [self.N * self.N, self.opt.cat_num])
 
-    def set_input(self, input):
-        """input: a dictionary that contains the data itself and its metadata information."""
-        self.input_imgs = input['image'].to(self.device)  
-        if self.opt.cgan:
-            self.input_targets = input['target'].to(self.device) 
-        
-    def forward(self, batch_size = None): 
+    def forward(self, batch_size=None):
         bs = self.opt.batch_size if batch_size is None else batch_size
-        if self.opt.z_type == 'Gaussian': 
-            z = torch.randn(bs, self.opt.z_dim, 1, 1, device=self.device) 
-        elif self.opt.z_type == 'Uniform': 
-            z = torch.rand(bs, self.opt.z_dim, 1, 1, device=self.device)*2. - 1. 
-        
+        if self.opt.z_type == 'Gaussian':
+            z = torch.randn(bs, self.opt.z_dim, 1, 1, device=self.device)
+        elif self.opt.z_type == 'Uniform':
+            z = torch.rand(bs, self.opt.z_dim, 1, 1, device=self.device) * 2. - 1.
+
         if not self.opt.cgan:
             self.gen_imgs = self.netG(z)
         else:
@@ -129,9 +122,9 @@ class TwoPlayerGANModel(BaseModel):
             self.fake_out = self.netD(self.gen_imgs, self.y_)
             self.real_out = self.netD(self.real_imgs, self.targets)
 
-        self.loss_G_fake, self.loss_G_real = self.criterionG(self.fake_out, self.real_out) 
+        self.loss_G_fake, self.loss_G_real = self.criterionG(self.fake_out, self.real_out)
         self.loss_G = self.loss_G_fake + self.loss_G_real
-        self.loss_G.backward() 
+        self.loss_G.backward()
 
     def backward_D(self):
         self.gen_imgs = self.gen_imgs.detach()
@@ -143,21 +136,23 @@ class TwoPlayerGANModel(BaseModel):
             self.fake_out = self.netD(self.gen_imgs, self.y_)
             self.real_out = self.netD(self.real_imgs, self.targets)
 
-        self.loss_D_fake, self.loss_D_real = self.criterionD(self.fake_out, self.real_out) 
-        if self.opt.use_gp is True: 
-            self.loss_D_gp = networks.cal_gradient_penalty(self.netD, self.real_imgs, self.gen_imgs, self.device, type='mixed', constant=1.0, lambda_gp=10.0)[0]
+        self.loss_D_fake, self.loss_D_real = self.criterionD(self.fake_out, self.real_out)
+        if self.opt.use_gp is True:
+            self.loss_D_gp = \
+            cal_gradient_penalty(self.netD, self.real_imgs, self.gen_imgs, self.device, type='mixed',
+                                          constant=1.0, lambda_gp=10.0)[0]
         else:
             self.loss_D_gp = 0.
 
         self.loss_D = self.loss_D_fake + self.loss_D_real + self.loss_D_gp
-        self.loss_D.backward() 
+        self.loss_D.backward()
 
     def optimize_parameters(self):
         for i in range(self.opt.D_iters + 1):
-            self.real_imgs = self.input_imgs[i*self.opt.batch_size:(i+1)*self.opt.batch_size,:,:,:]
+            self.real_imgs = self.input_imgs[i * self.opt.batch_size:(i + 1) * self.opt.batch_size, :, :, :]
             if self.opt.cgan:
-                self.targets = self.input_target[i*self.opt.batch_size:(i+1)*self.opt.batch_size,:] 
-            self.forward()               
+                self.targets = self.input_target[i * self.opt.batch_size:(i + 1) * self.opt.batch_size, :]
+            self.forward()
             # update G
             if i == 0:
                 self.set_requires_grad(self.netD, False)
@@ -165,9 +160,8 @@ class TwoPlayerGANModel(BaseModel):
                 self.backward_G()
                 self.optimizer_G.step()
             # update D
-            else: 
+            else:
                 self.set_requires_grad(self.netD, True)
                 self.optimizer_D.zero_grad()
                 self.backward_D()
                 self.optimizer_D.step()
-
